Ft chats

+++ server/src/messages/messages.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, Logger } from '@nestjs/common';
+import { ConflictException, Injectable, Logger } from '@nestjs/common';
 import { MessagesRepository } from './messages.repository';
 import { SendPrivateMessageDto } from './dto/send-private-message.dto';
 import { SendPublicMessageDto } from './dto/send-public-message.dto';
@@ -6,6 +6,14 @@ import { GetPrivateChatMessagesDto } from './dto/get-private-chat-messages.dto';
 import { FindMessageDto } from './dto/find-message.dto';
 import { UpdateMessageDto } from './dto/update-message.dto';
 import { UsersService } from 'src/users/users.service';
+import { User } from 'src/users/models/user.schema';
+
+interface MessageResponseDto {
+  _id: string;
+  text: string;
+  sender: { _id: string; username: string };
+  receiver?: { _id: string; username: string };
+}
 
 @Injectable()
 export class MessagesService {
@@ -54,21 +62,36 @@ export class MessagesService {
     senderId,
     receiverId,
     text,
-  }: SendPrivateMessageDto) {
+  }: SendPrivateMessageDto): Promise<MessageResponseDto> {
     try {
-      return await this.messagesRepository.create({
+      const messageCreated = await this.messagesRepository.create({
         sender: senderId,
         receiver: receiverId,
         text,
-        populate: ['sender', 'receiver'],
       });
+
+      const sender = await this.usersService.getUser({ _id: senderId });
+      const receiver = await this.usersService.getUser({ _id: receiverId });
+
+      return {
+        _id: messageCreated._id.toHexString(),
+        text: messageCreated.text,
+        sender: { _id: sender._id.toHexString(), username: sender.username },
+        receiver: {
+          _id: receiver._id.toHexString(),
+          username: receiver.username,
+        },
+      };
     } catch (err) {
       this.logger.error(err);
       throw err;
     }
   }
 
-  async sendPublicMessage({ senderId, text }: SendPublicMessageDto) {
+  async sendPublicMessage({
+    senderId,
+    text,
+  }: SendPublicMessageDto): Promise<MessageResponseDto> {
     try {
       const sender = await this.usersService.findById(senderId);
 
@@ -80,7 +103,6 @@ export class MessagesService {
         sender: senderId,
         text,
         receiver: null,
-        populate: ['sender'],
       });
 
       return {
@@ -89,8 +111,6 @@ export class MessagesService {
         sender: {
           _id: sender._id.toHexString(),
           username: sender.username,
-          email: sender.email,
-          avatar: sender.avatar,
         },
       };
     } catch (err) {
@@ -99,24 +119,71 @@ export class MessagesService {
     }
   }
 
-  async deleteMessage({ messageId, userId }: FindMessageDto) {
+  async deleteMessage({ messageId, userId }: FindMessageDto): Promise<string> {
     try {
-      return await this.messagesRepository.findOneAndDelete({
+      const message = await this.messagesRepository.findOne({
         _id: messageId,
         sender: userId,
       });
+
+      if (!message) {
+        throw new ConflictException(
+          'You are not the sender or message does not exists',
+        );
+      }
+
+      await this.messagesRepository.findOneAndDelete({
+        _id: messageId,
+        sender: userId,
+      });
+
+      return message._id.toHexString();
     } catch (err) {
       this.logger.error(err);
       throw err;
     }
   }
 
-  async updateMessage({ messageId, userId, text }: UpdateMessageDto) {
+  async updateMessage({
+    messageId,
+    userId,
+    text,
+  }: UpdateMessageDto): Promise<MessageResponseDto> {
     try {
-      return await this.messagesRepository.findOneAndUpdate(
+      const updatedMessage = await this.messagesRepository.findOneAndUpdate(
         { sender: userId, _id: messageId },
-        { text, new: true, populate: ['sender'] },
+        { text, new: true },
       );
+
+      if (!updatedMessage) {
+        throw new ConflictException(
+          'You are not the sender or message does not exists',
+        );
+      }
+
+      const sender = await this.usersService.getUser({
+        _id: userId,
+      });
+
+      let receiver: User | null = null;
+
+      if (updatedMessage.receiver?._id) {
+        receiver = await this.usersService.getUser({
+          _id: updatedMessage.receiver?.toHexString?.(),
+        });
+      }
+
+      return {
+        _id: updatedMessage._id.toHexString(),
+        text: updatedMessage.text,
+        sender: { _id: sender._id.toHexString(), username: sender.username },
+        receiver: receiver
+          ? {
+              _id: receiver?._id.toHexString(),
+              username: receiver?.username,
+            }
+          : undefined,
+      };
     } catch (err) {
       this.logger.error(err);
       throw err;
